pyproject.toml
[tool.poetry]
name = "alliumassignment"
version = "0.1.0"
description = ""
authors = ["eugeneleejunping <eugeneleejunping@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
retry = "^0.9.2"
aiohttp = "^3.11.8"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.2"
typing-extensions = "^4.12.2"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
dbt-core = "^1.8.9"
dbt-postgres = "^1.8.2"
types-retry = "^0.9.9.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"
mypy = "^1.13.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
exclude = '^(\\.venv|venv|env|\\.git|node_modules)/'